Add list to show the entries of a result list
This list is supposed to show the entries of the query
which is triggered by the user input into the search bar.
It uses a data grid component from mui. I removed header
and footer and allowed only one page of pagination for
styling causes. The list is added to the parent Add.js
below the search bar.