Add fetchService and handler which triggers a fetch on search
A fetch service is added to fetch data. A mockFetch function is
added to simulate fetching data from a server. In the main component
Add.js a handler is added which triggers the fetching based upon the
input in the search bar. Is passed to the search bar component.
Furthermore, the telefonuch.json is added to src/assets in case
i want to use it for testing purposes later.