Add css transitions for better user experience
When results are shown in the list, the search bar
moves up. With transitions, this is happening much
smoother now, whith a bit of a back bounce. Furthermore,
when the list is shown, the lower borders of the search
bar drop the radius (are not round anymore). This is done
for design reasons so the search bar 'connects' with the list.